fix(ci): python-coverage-shards — split two test-part files back under the 1500-line maintainability limit
test_pr_monitor_runner_part_005.py (1543) and
test_pr_monitor_runner_coverage_edges_part_003.py (1518) exceeded the
MAX_FIRST_PARTY_FILE_LINES=1500 guardrail enforced by
test_first_party_code_files_stay_under_line_limit, failing the coverage
shard. Move whole, self-contained test functions into sibling part files
with headroom (part_007 and part_004), drop the now-unused imports left
behind, and add only the imports each moved test needs in its new home.
No test behavior changed.
