+++ tests/unit/runtime/test_pr_monitor_runner_coverage_edges_parts/test_pr_monitor_runner_coverage_edges_part_004.py
@@ -7,10 +7,11 @@ from pathlib import Path
 from typing import Any
 
 import pytest
+import structlog
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
 from awf.common.bitbucket_client import BitbucketClientError
-from awf.common.commands import FakeCommandRunner
+from awf.common.commands import CommandResult, FakeCommandRunner
 from awf.common.github_client import RepoRef
 from awf.db.enums import OperationStatus, WorkspaceStatus
 from awf.db.repositories import (
@@ -36,6 +37,7 @@ from awf.runtime.pr_monitor import (
 )
 from awf.runtime.pr_monitor_runner.helpers import (
     _review_comment_body_state_key,
+    _target_reconcile_failure_payload,
 )
 from awf.runtime.pr_monitor_runner.remote_ops import (
     _GitPushResult,
@@ -43,6 +45,7 @@ from awf.runtime.pr_monitor_runner.remote_ops import (
 from awf.runtime.pr_monitor_runner.types import (
     _MonitorAgentRuntimeOwnershipRepairFailedError,
 )
+from awf.service.target_branch_monitor import TargetBranchMonitorError
 from tests.postgres import postgres_test_engine
 from tests.shared.monitor_runner import DefaultMergeMethodGitHubClient
 from tests.unit.runtime._monitor_runner_fixtures import (
@@ -1052,3 +1055,138 @@ async def test_sync_base_conflict_ownership_repair_failure_blocks_push(
     assert push_result.stderr == AGENT_RUNTIME_OWNERSHIP_REPAIR_FAILED_REASON_CODE
     assert len(adapter.calls) == 1
     assert not any(call.args[:1] == ["git"] and "push" in call.args for call in cmd.calls)
+
+
+@pytest.mark.unit
+async def test_target_branch_reconcile_failure_reuses_exception_payload(
+    factory: async_sessionmaker[AsyncSession],
+    tmp_path: Path,
+) -> None:
+    workspace_id = await seed_monitoring_workspace(factory)
+
+    class _CountingResultError(Exception):
+        def __init__(self) -> None:
+            super().__init__("target branch locked " + ("x" * 1200))
+            self.result_accesses = 0
+            self._result = CommandResult(
+                returncode=128,
+                stdout="stdout " + ("o" * 1200),
+                stderr="stderr " + ("e" * 1200),
+                reason_code="GIT_FAILED",
+            )
+
+        @property
+        def result(self) -> CommandResult:
+            self.result_accesses += 1
+            return self._result
+
+    failure = _CountingResultError()
+
+    async def failing_reconciler(*, repo_url: str, branch: str, workspace_id: str) -> object:
+        del repo_url, branch, workspace_id
+        raise failure
+
+    runner = make_runner(
+        factory=factory,
+        cmd=FakeCommandRunner(),
+        adapter=FakeAdapter(),
+        sleep_fn=RecordedSleep(),
+        worktrees_root=tmp_path / "worktrees",
+        post_merge_target_reconciler=failing_reconciler,
+    )
+
+    with structlog.testing.capture_logs() as captured:
+        await runner._reconcile_target_branch_after_merge(
+            workspace_id=workspace_id,
+            repo_url="git@github.com:dimileeh/aira-web.git",
+            base_branch="development",
+        )
+
+    failure_log = next(
+        event
+        for event in captured
+        if event.get("event") == "monitor.target_branch_reconcile_failed"
+    )
+    assert failure.result_accesses == 1
+    assert len(failure_log["error"]) == 500
+    assert len(failure_log["stderr"]) == 500
+    assert len(failure_log["stdout"]) == 500
+
+    async with factory() as s:
+        ws = await WorkspaceRepository(s).get(workspace_id)
+        assert ws is not None
+        event_payload = ws.events[-1].payload
+        assert len(event_payload["error"]) == 1000
+        assert len(event_payload["stderr"]) == 1000
+        assert len(event_payload["stdout"]) == 1000
+
+
+@pytest.mark.unit
+def test_target_reconcile_failure_payload_uses_command_error_contract() -> None:
+    result = CommandResult(
+        returncode=128,
+        stdout="fatal stdout",
+        stderr="fatal stderr",
+        reason_code="GIT_FAILED",
+    )
+    exc = TargetBranchMonitorError(
+        operation="target_branch.git_fetch",
+        result=result,
+    )
+    exc.target_reconcile_payload = lambda: {  # type: ignore[attr-defined]
+        "status": "committed",
+        "resolver_results": [{"status": "resolved"}],
+        "commit_sha": "abc123",
+        "pushed": True,
+    }
+
+    payload = _target_reconcile_failure_payload(exc, error_limit=100)
+
+    assert payload["status"] == "failed"
+    assert "target_reconcile_status" not in payload
+    assert payload["resolver_results"] == []
+    assert payload["commit_sha"] is None
+    assert payload["pushed"] is False
+    assert payload["operation"] == "target_branch.git_fetch"
+    assert payload["returncode"] == 128
+    assert payload["command_reason_code"] == "GIT_FAILED"
+    assert payload["stderr"] == "fatal stderr"
+    assert payload["stdout"] == "fatal stdout"
+
+
+@pytest.mark.unit
+async def test_target_branch_reconcile_success_appends_payload_event(
+    factory: async_sessionmaker[AsyncSession],
+    tmp_path: Path,
+) -> None:
+    workspace_id = await seed_monitoring_workspace(factory)
+
+    async def reconciler(*, repo_url: str, branch: str, workspace_id: str) -> object:
+        return {
+            "status": "TARGET_BRANCH_FAST_FORWARDED",
+            "repo_url": repo_url,
+            "branch": branch,
+            "workspace_id": workspace_id,
+        }
+
+    runner = make_runner(
+        factory=factory,
+        cmd=FakeCommandRunner(),
+        adapter=FakeAdapter(),
+        sleep_fn=RecordedSleep(),
+        worktrees_root=tmp_path / "worktrees",
+        post_merge_target_reconciler=reconciler,
+    )
+
+    await runner._reconcile_target_branch_after_merge(
+        workspace_id=workspace_id,
+        repo_url="git@github.com:dimileeh/aira-web.git",
+        base_branch="development",
+    )
+
+    async with factory() as s:
+        ws = await WorkspaceRepository(s).get(workspace_id)
+        assert ws is not None
+        assert ws.events[-1].event_type == "target_branch.reconciled"
+        assert ws.events[-1].reason_code == "TARGET_BRANCH_FAST_FORWARDED"
+        assert ws.events[-1].payload["branch"] == "development"
