+++ tests/unit/runtime/test_pr_monitor_runner_parts/test_pr_monitor_runner_part_005.py
@@ -47,7 +47,6 @@ from awf.runtime.pr_monitor import (
     MergeableState,
     MergeStateStatus,
     MonitorState,
-    NotifyHuman,
     PRStatus,
     ReviewComment,
 )
@@ -1181,128 +1180,6 @@ async def test_late_validation_recovery_callback_records_stale_ready_workspace(
     assert stale_events[0].payload["callback_action"] == "recovery_dispatch"
     assert stale_events[0].payload["actual_status"] == WorkspaceStatus.ready.value
     assert [op for op in operations if op.type == OperationType.validate.value] == []
-
-
-@pytest.mark.unit
-async def test_manual_human_wait_records_operation(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    cmd = FakeCommandRunner()
-    cmd.queue_result(returncode=0)
-    sleep_fn = RecordedSleep()
-    pr_number = 79
-    head_sha = "f" * 40
-    workspace_id = await seed_monitoring_workspace(
-        factory,
-        pr_number=pr_number,
-        head_sha=head_sha,
-        auto_merge=False,
-    )
-    runner = make_runner(
-        factory=factory,
-        cmd=cmd,
-        adapter=FakeAdapter(),
-        sleep_fn=sleep_fn,
-        worktrees_root=tmp_path / "worktrees",
-    )
-
-    terminal = await runner._execute(
-        action=NotifyHuman(message="manual merge required"),
-        workspace_id=workspace_id,
-        repo_url="git@github.com:dimileeh/aira-web.git",
-        repo=RepoRef(owner="dimileeh", name="aira-web"),
-        pr_number=pr_number,
-        status=_green_status(pr_number=pr_number, head_sha=head_sha),
-        state=MonitorState(started_at=0.0),
-        base_branch="development",
-        remote_branch=f"awf/{workspace_id}",
-        compose_project="proj",
-        compose_file=tmp_path / "compose.yml",
-        monitor_log=None,
-    )
-
-    assert terminal is False
-    assert sleep_fn.calls == [60]
-    async with factory() as session:
-        workspace = await WorkspaceRepository(session).get(workspace_id)
-        assert workspace is not None
-        operations = await OperationRepository(session).list_all(workspace_id=workspace_id)
-    assert workspace.status == WorkspaceStatus.monitoring_pr.value
-    assert len(operations) == 1
-    operation = operations[0]
-    assert operation.type == "human_wait"
-    assert operation.status == OperationStatus.succeeded.value
-    assert operation.started_at is not None
-    assert operation.finished_at is not None
-    assert operation.payload == {
-        "owner": "pr_monitor",
-        "source": "pr_monitor",
-        "action": "human_wait",
-        "requested_action": "notify_human",
-        "reason": "manual merge required",
-        "reason_code": "HUMAN_WAIT",
-        "pr_number": pr_number,
-        "pr_url": f"https://github.com/dimileeh/aira-web/pull/{pr_number}",
-        "source_head_sha": head_sha,
-        "source_base_sha": "a" * 40,
-        "target_branch": "development",
-        "remote_branch": f"awf/{workspace_id}",
-    }
-    assert operation.result == {
-        "status": "succeeded",
-        "outcome": "human_notification_posted",
-        "slept_seconds": 60,
-    }
-
-
-@pytest.mark.unit
-async def test_monitor_operation_payload_redacts_secret_like_values(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path: Path,
-) -> None:
-    cmd = FakeCommandRunner()
-    cmd.queue_result(returncode=0)
-    workspace_id = await seed_monitoring_workspace(factory, pr_number=80)
-    runner = make_runner(
-        factory=factory,
-        cmd=cmd,
-        adapter=FakeAdapter(),
-        sleep_fn=RecordedSleep(),
-        worktrees_root=tmp_path / "worktrees",
-    )
-
-    await runner._execute(
-        action=NotifyHuman(
-            message=(
-                "blocked with Bearer ghp_should_not_persist "
-                "token=github_pat_should_not_persist password=sk-should-not-persist"
-            )
-        ),
-        workspace_id=workspace_id,
-        repo_url="git@github.com:dimileeh/aira-web.git",
-        repo=RepoRef(owner="dimileeh", name="aira-web"),
-        pr_number=80,
-        status=_green_status(pr_number=80),
-        state=MonitorState(started_at=0.0),
-        base_branch="development",
-        remote_branch=f"awf/{workspace_id}",
-        compose_project="proj",
-        compose_file=tmp_path / "compose.yml",
-        monitor_log=None,
-    )
-
-    async with factory() as session:
-        operations = await OperationRepository(session).list_all(workspace_id=workspace_id)
-    assert len(operations) == 1
-    persisted = f"{operations[0].payload!r} {operations[0].result!r}"
-    assert "ghp_should_not_persist" not in persisted
-    assert "github_pat_should_not_persist" not in persisted
-    assert "sk-should-not-persist" not in persisted
-    assert "Bearer" not in persisted
-    assert "token=" not in persisted
-    assert "password=" not in persisted
-    assert "[redacted]" in persisted
 
 
 @pytest.mark.unit
